build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(lwio: r39716)

+++ lwio/server/pvfs/string.c
@@ -62,7 +62,7 @@ PvfsCStringUpper(
     PSTR p = pszString;
 
     while (PVFS_CSTRING_NON_NULL(p)) {
-        *p = toupper(*p);
+        *p = toupper((int)*p);
         p++;
     }
